requirements.txt
aiofiles==24.1.0
alembic==1.13.3
celery==5.5.3
click==8.2.1
cos_python_sdk_v5==1.9.37
fastembed==0.7.1
Flask==3.1.1
flask_cors==6.0.1
Flask_Login==0.6.3
Flask_Migrate==4.1.0
flask_sqlalchemy==3.1.1
flask_wtf==1.2.2
injector==0.21.0
jieba==0.42.1
langchain==0.3.26
langchain_community==0.3.27
langchain_core==0.3.68
langchain_huggingface==0.3.0
langchain_openai==0.3.27
langchain_text_splitters==0.3.8
langchain_weaviate==0.0.5
marshmallow==4.0.0
numpy==2.3.1
openai==1.95.1
pydantic==2.11.7
PyJWT==2.10.1
PyJWT==2.10.1
python-dotenv==1.1.1
PyYAML==6.0.2
PyYAML==6.0.2
redis==6.2.0
Requests==2.32.4
weaviate==0.1.2
weaviate_client==4.15.4
Werkzeug==3.1.3
WTForms==3.2.1
